+++ tests/utils.py
@@ -1,3 +1,7 @@
+import contextlib
+import socket
+import subprocess
+import time
 from typing import Tuple
 
 import sniffio
@@ -9,6 +13,17 @@ def lookup_async_backend():
 
 def lookup_sync_backend():
     return "sync"
+
+
+def wait_until_pproxy_serve_on_port(host: str, port: int):
+    while True:
+        try:
+            sock = socket.create_connection((host, port))
+        except ConnectionRefusedError:
+            time.sleep(0.25)
+        else:
+            sock.close()
+            break
 
 
 class Server:
@@ -27,3 +42,19 @@ class Server:
     @property
     def host_header(self) -> Tuple[bytes, bytes]:
         return (b"host", self._host.encode("utf-8"))
+
+
+@contextlib.contextmanager
+def http_proxy_server(proxy_host: str, proxy_port: int):
+
+    proc = None
+    try:
+        command = ["pproxy", "-l", f"http://{proxy_host}:{proxy_port}/"]
+        proc = subprocess.Popen(command)
+
+        wait_until_pproxy_serve_on_port(proxy_host, proxy_port)
+
+        yield b"http", proxy_host.encode(), proxy_port, b"/"
+    finally:
+        if proc is not None:
+            proc.kill()
